graphQl01/requirements.txt
django
graphene-django
confluent_kafka
celery
psycopg2-binary
django-extensions
python-decouple